LPS-64093 extract for simplicity, support jsp file (note this class has to be in this module)
